Renamed Project.h/cpp to HoI4Project.h/cpp.
